first set of suggestions from Isaac
